move filter specs to filters package
